add clean architecture
