+++ src/core/services/authentication.service.ts
@@ -1,14 +1,15 @@
 import { Injectable } from '@angular/core';
 import { ApiService } from './api.service';
 import { HttpClient } from '@angular/common/http';
-import { CreateRespondentsAccountsDto } from '../models/create.respondents.accounts.dto';
 import { Observable } from 'rxjs';
-import { LoginDto } from '../models/login.dto';
+import { AuthenticationService } from '../../domain/external_services/authentication.service';
+import { CreateRespondentsAccountsDto } from '../../domain/models/create.respondents.accounts.dto';
+import { LoginDto } from '../../domain/models/login.dto';
 
 @Injectable({
   providedIn: 'root'
 })
-export class AuthenticationService  extends ApiService{
+export class AuthenticationServiceImpl  extends ApiService implements AuthenticationService {
 
   constructor(client: HttpClient) {
     super(client);
