+++ src/domain/external_services/authentication.service.ts
@@ -1,0 +1,7 @@
+import { Observable } from "rxjs";
+import { CreateRespondentsAccountsDto } from "../models/create.respondents.accounts.dto";
+import { LoginDto } from "../models/login.dto";
+
+export interface AuthenticationService{
+    generateRespondents(dto: CreateRespondentsAccountsDto): Observable<LoginDto[]>;
+}
