+++ src/presentation/modules/app/components/add-respondents/add-respondents.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { CreateRespondentsAccountsDto } from '../core/models/create.respondents.accounts.dto';
 import { FormControl, Validators, FormGroup, FormBuilder } from '@angular/forms';
-import { AuthenticationService } from '../core/services/authentication.service';
-import { LoginDto } from '../core/models/login.dto';
 import { Clipboard } from '@angular/cdk/clipboard';
+import { AuthenticationServiceImpl } from '../../../../../core/services/authentication.service';
+import { LoginDto } from '../../../../../domain/models/login.dto';
+import { CreateRespondentsAccountsDto } from '../../../../../domain/models/create.respondents.accounts.dto';
 
 interface FormGroupType{
   amount: FormControl<number | null>
@@ -24,7 +24,7 @@ export class AddRespondentsComponent implements OnInit{
   };
 
   constructor(private formBuilder: FormBuilder,
-    private readonly _service: AuthenticationService,
+    private readonly _service: AuthenticationServiceImpl,
     private readonly _clipboard: Clipboard){
   }
