+++ src/presentation/modules/app/components/app/app.component.html
@@ -1,0 +1,2 @@
+<router-outlet *ngIf="!shouldDisplayDashboard()"/>
+<app-dashboard *ngIf="shouldDisplayDashboard()">
