+++ src/presentation/modules/app/components/dashboard/dashboard.component.html
@@ -15,6 +15,6 @@
         </mat-nav-list>
   </mat-drawer>
   <mat-drawer-content>
-      <app-respondents/>
+      <router-outlet/>
   </mat-drawer-content>
 </mat-drawer-container>
