+++ src/presentation/modules/app/components/login/login.component.ts
@@ -26,6 +26,6 @@ export class LoginComponent {
   constructor(private readonly router: Router){}
 
   submit(){
-    this.router.navigate(['/dashboard']);
+    this.router.navigate(['/']);
   }
 }
